+++ src/factories/plugin-factory.ts
@@ -17,6 +17,7 @@ import {
   PluginType,
   RepositoryConfig,
   SentenceCasePluginConfig,
+  GroupPriorityPluginConfig,
 } from '../manifest';
 import {GitHub} from '../github';
 import {ManifestPlugin} from '../plugin';
@@ -27,6 +28,9 @@ import {VersioningStrategyType} from './versioning-strategy-factory';
 import {MavenWorkspace} from '../plugins/maven-workspace';
 import {ConfigurationError} from '../errors';
 import {SentenceCase} from '../plugins/sentence-case';
+import {GroupPriority} from '../plugins/group-priority';
+import {Logger} from '../util/logger';
+import {WorkspacePluginOptions} from '../plugins/workspace';
 
 export interface PluginFactoryOptions {
   type: PluginType;
@@ -40,6 +44,9 @@ export interface PluginFactoryOptions {
 
   // workspace options
   updateAllPackages?: boolean;
+  considerAllArtifacts?: boolean;
+
+  logger?: Logger;
 }
 
 export type PluginBuilder = (options: PluginFactoryOptions) => ManifestPlugin;
@@ -58,21 +65,30 @@ const pluginFactories: Record<string, PluginBuilder> = {
       options.github,
       options.targetBranch,
       options.repositoryConfig,
-      options
+      {
+        ...options,
+        ...(options.type as WorkspacePluginOptions),
+      }
     ),
   'node-workspace': options =>
     new NodeWorkspace(
       options.github,
       options.targetBranch,
       options.repositoryConfig,
-      options
+      {
+        ...options,
+        ...(options.type as WorkspacePluginOptions),
+      }
     ),
   'maven-workspace': options =>
     new MavenWorkspace(
       options.github,
       options.targetBranch,
       options.repositoryConfig,
-      options
+      {
+        ...options,
+        ...(options.type as WorkspacePluginOptions),
+      }
     ),
   'sentence-case': options =>
     new SentenceCase(
@@ -80,6 +96,13 @@ const pluginFactories: Record<string, PluginBuilder> = {
       options.targetBranch,
       options.repositoryConfig,
       (options.type as SentenceCasePluginConfig).specialWords
+    ),
+  'group-priority': options =>
+    new GroupPriority(
+      options.github,
+      options.targetBranch,
+      options.repositoryConfig,
+      (options.type as GroupPriorityPluginConfig).groups
     ),
 };