+++ src/plugins/workspace.ts
@@ -104,47 +104,49 @@ export abstract class WorkspacePlugin<T> extends ManifestPlugin {
     );
     const graph = await this.buildGraph(allPackages);
 
-    const packageNamesToUpdate = this.updateAllPackages
-      ? allPackages.map(this.packageNameFromPackage)
-      : Object.keys(candidatesByPackage);
+    const packageNamesToUpdate = this.packageNamesToUpdate(
+      graph,
+      candidatesByPackage
+    );
     const orderedPackages = this.buildGraphOrder(graph, packageNamesToUpdate);
     this.logger.info(`Updating ${orderedPackages.length} packages`);
 
-    const updatedVersions: VersionsMap = new Map();
-    const updatedPathVersions: VersionsMap = new Map();
-    for (const pkg of orderedPackages) {
-      const packageName = this.packageNameFromPackage(pkg);
-      this.logger.debug(`package: ${packageName}`);
-      const existingCandidate = candidatesByPackage[packageName];
-      if (existingCandidate) {
-        const version = existingCandidate.pullRequest.version!;
-        this.logger.debug(`version: ${version} from release-please`);
-        updatedVersions.set(packageName, version);
-      } else {
-        const version = this.bumpVersion(pkg);
-        this.logger.debug(`version: ${version} forced bump`);
-        updatedVersions.set(packageName, version);
-        updatedPathVersions.set(this.pathFromPackage(pkg), version);
-      }
-    }
+    const {updatedVersions, updatedPathVersions} =
+      await this.buildUpdatedVersions(
+        graph,
+        orderedPackages,
+        candidatesByPackage
+      );
 
     let newCandidates: CandidateReleasePullRequest[] = [];
+    // In some cases, there are multiple packages within a single candidate. We
+    // only want to process each candidate package once.
+    const newCandidatePaths = new Set<string>();
     for (const pkg of orderedPackages) {
-      const packageName = this.packageNameFromPackage(pkg);
-      const existingCandidate = candidatesByPackage[packageName];
+      const existingCandidate = this.findCandidateForPackage(
+        pkg,
+        candidatesByPackage
+      );
       if (existingCandidate) {
         // if already has an pull request, update the changelog and update
         this.logger.info(
           `Updating exising candidate pull request for ${this.packageNameFromPackage(
             pkg
-          )}`
+          )}, path: ${existingCandidate.path}`
         );
-        const newCandidate = this.updateCandidate(
-          existingCandidate,
-          pkg,
-          updatedVersions
-        );
-        newCandidates.push(newCandidate);
+        if (newCandidatePaths.has(existingCandidate.path)) {
+          this.logger.info(
+            `Already updated candidate for path: ${existingCandidate.path}`
+          );
+        } else {
+          const newCandidate = this.updateCandidate(
+            existingCandidate,
+            pkg,
+            updatedVersions
+          );
+          newCandidatePaths.add(newCandidate.path);
+          newCandidates.push(newCandidate);
+        }
       } else {
         // otherwise, build a new pull request with changelog and entry update
         this.logger.info(
@@ -153,7 +155,14 @@ export abstract class WorkspacePlugin<T> extends ManifestPlugin {
           )}`
         );
         const newCandidate = this.newCandidate(pkg, updatedVersions);
-        newCandidates.push(newCandidate);
+        if (newCandidatePaths.has(newCandidate.path)) {
+          this.logger.info(
+            `Already created new candidate for path: ${newCandidate.path}`
+          );
+        } else {
+          newCandidatePaths.add(newCandidate.path);
+          newCandidates.push(newCandidate);
+        }
       }
     }
 
@@ -183,6 +192,87 @@ export abstract class WorkspacePlugin<T> extends ManifestPlugin {
     newCandidates = this.postProcessCandidates(newCandidates, updatedVersions);
 
     return [...outOfScopeCandidates, ...newCandidates];
+  }
+
+  /**
+   * Helper for finding a candidate release based on the package name.
+   * By default, we assume that the package name matches the release
+   * component.
+   * @param {T} pkg The package being released
+   * @param {Record<string, CandidateReleasePullRequest} candidatesByPackage
+   *   The candidate pull requests indexed by the package name.
+   * @returns {CandidateReleasePullRequest | undefined} The associated
+   *   candidate release or undefined if there is no existing release yet
+   */
+  protected findCandidateForPackage(
+    pkg: T,
+    candidatesByPackage: Record<string, CandidateReleasePullRequest>
+  ): CandidateReleasePullRequest | undefined {
+    const packageName = this.packageNameFromPackage(pkg);
+    return candidatesByPackage[packageName];
+  }
+
+  /**
+   * Helper to determine which packages we will use to base our search
+   * for touched packages upon. These are usually the packages that
+   * have candidate pull requests open.
+   *
+   * If you configure `updateAllPackages`, we fill force update all
+   * packages as if they had a release.
+   * @param {DependencyGraph<T>} graph All the packages in the repository
+   * @param {Record<string, CandidateReleasePullRequest} candidatesByPackage
+   *   The candidate pull requests indexed by the package name.
+   * @returns {string[]} Package names to
+   */
+  protected packageNamesToUpdate(
+    graph: DependencyGraph<T>,
+    candidatesByPackage: Record<string, CandidateReleasePullRequest>
+  ): string[] {
+    if (this.updateAllPackages) {
+      return Array.from(graph.values()).map(({value}) =>
+        this.packageNameFromPackage(value)
+      );
+    }
+    return Object.keys(candidatesByPackage);
+  }
+
+  /**
+   * Helper to build up all the versions we are modifying in this
+   * repository.
+   * @param {DependencyGraph<T>} _graph All the packages in the repository
+   * @param {T[]} orderedPackages A list of packages that are currently
+   *   updated by the existing candidate pull requests
+   * @param {Record<string, CandidateReleasePullRequest} candidatesByPackage
+   *   The candidate pull requests indexed by the package name.
+   * @returns A map of all updated versions (package name => Version) and a
+   *   map of all updated versions (component path => Version).
+   */
+  protected async buildUpdatedVersions(
+    _graph: DependencyGraph<T>,
+    orderedPackages: T[],
+    candidatesByPackage: Record<string, CandidateReleasePullRequest>
+  ): Promise<{updatedVersions: VersionsMap; updatedPathVersions: VersionsMap}> {
+    const updatedVersions: VersionsMap = new Map();
+    const updatedPathVersions: VersionsMap = new Map();
+    for (const pkg of orderedPackages) {
+      const packageName = this.packageNameFromPackage(pkg);
+      this.logger.debug(`package: ${packageName}`);
+      const existingCandidate = candidatesByPackage[packageName];
+      if (existingCandidate) {
+        const version = existingCandidate.pullRequest.version!;
+        this.logger.debug(`version: ${version} from release-please`);
+        updatedVersions.set(packageName, version);
+      } else {
+        const version = this.bumpVersion(pkg);
+        this.logger.debug(`version: ${version} forced bump`);
+        updatedVersions.set(packageName, version);
+        updatedPathVersions.set(this.pathFromPackage(pkg), version);
+      }
+    }
+    return {
+      updatedVersions,
+      updatedPathVersions,
+    };
   }
 
   /**
@@ -344,6 +434,7 @@ export abstract class WorkspacePlugin<T> extends ManifestPlugin {
     visited: Set<T>,
     path: string[]
   ) {
+    this.logger.debug(`visiting ${name}, path: ${path}`);
     if (path.indexOf(name) !== -1) {
       throw new Error(
         `found cycle in dependency graph: ${path.join(' -> ')} -> ${name}`
@@ -363,6 +454,7 @@ export abstract class WorkspacePlugin<T> extends ManifestPlugin {
         this.logger.warn(`dependency not found in graph: ${depName}`);
         return;
       }
+      this.logger.info(`visiting ${depName} next`);
 
       this.visitPostOrder(graph, depName, visited, nextPath);
     }
@@ -374,6 +466,8 @@ export abstract class WorkspacePlugin<T> extends ManifestPlugin {
         )} to order`
       );
       visited.add(node.value);
+    } else {
+      this.logger.debug(`${node.value} already visited`);
     }
   }
 }