+++ test/plugins/maven-workspace.ts
@@ -23,11 +23,14 @@ import {
   buildGitHubFileContent,
   stubFilesFromFixtures,
   safeSnapshot,
+  assertHasUpdate,
+  assertHasUpdates,
 } from '../helpers';
 import {expect} from 'chai';
 import {Update} from '../../src/update';
 import {Version} from '../../src/version';
 import {PomXml} from '../../src/updaters/java/pom-xml';
+import {RawContent} from '../../src/updaters/raw-content';
 
 const sandbox = sinon.createSandbox();
 const fixturesPath = './test/fixtures/plugins/maven-workspace';
@@ -42,6 +45,9 @@ describe('MavenWorkspace plugin', () => {
       defaultBranch: 'main',
     });
     plugin = new MavenWorkspace(github, 'main', {
+      bom: {
+        releaseType: 'maven',
+      },
       maven1: {
         releaseType: 'maven',
       },
@@ -62,9 +68,12 @@ describe('MavenWorkspace plugin', () => {
   describe('run', () => {
     it('handles a single maven package', async () => {
       const candidates: CandidateReleasePullRequest[] = [
-        buildMockCandidatePullRequest('maven4', 'maven', '4.4.5', 'maven4', [
-          buildMockPackageUpdate('maven4/pom.xml', 'maven4/pom.xml', '4.4.5'),
-        ]),
+        buildMockCandidatePullRequest('maven4', 'maven', '4.4.5', {
+          component: 'maven4',
+          updates: [
+            buildMockPackageUpdate('maven4/pom.xml', 'maven4/pom.xml', '4.4.5'),
+          ],
+        }),
       ];
       stubFilesFromFixtures({
         sandbox,
@@ -95,17 +104,19 @@ describe('MavenWorkspace plugin', () => {
     });
     it('appends to existing candidate', async () => {
       const candidates: CandidateReleasePullRequest[] = [
-        buildMockCandidatePullRequest('maven3', 'maven', '3.3.4', 'maven3', [
-          buildMockPackageUpdate('maven3/pom.xml', 'maven3/pom.xml', '3.3.4'),
-        ]),
-        buildMockCandidatePullRequest(
-          'maven4',
-          'maven',
-          '4.4.5',
-          'maven4',
-          [buildMockPackageUpdate('maven4/pom.xml', 'maven4/pom.xml', '4.4.5')],
-          '### Dependencies\n\n* Updated foo to v3'
-        ),
+        buildMockCandidatePullRequest('maven3', 'maven', '3.3.4', {
+          component: 'maven3',
+          updates: [
+            buildMockPackageUpdate('maven3/pom.xml', 'maven3/pom.xml', '3.3.4'),
+          ],
+        }),
+        buildMockCandidatePullRequest('maven4', 'maven', '4.4.5', {
+          component: 'maven4',
+          updates: [
+            buildMockPackageUpdate('maven4/pom.xml', 'maven4/pom.xml', '4.4.5'),
+          ],
+          notes: '### Dependencies\n\n* Updated foo to v3',
+        }),
       ];
       stubFilesFromFixtures({
         sandbox,
@@ -134,11 +145,78 @@ describe('MavenWorkspace plugin', () => {
       safeSnapshot(newCandidates[0].pullRequest.body.toString());
       expect(newCandidates[0].pullRequest.body.releaseData).length(2);
     });
+    it('appends to existing candidate with special updater', async () => {
+      const customUpdater = new RawContent('some content');
+      const candidates: CandidateReleasePullRequest[] = [
+        buildMockCandidatePullRequest('maven3', 'maven', '3.3.4', {
+          component: 'maven3',
+          updates: [
+            buildMockPackageUpdate('maven3/pom.xml', 'maven3/pom.xml', '3.3.4'),
+          ],
+        }),
+        buildMockCandidatePullRequest('maven4', 'maven', '4.4.5', {
+          component: 'maven4',
+          updates: [
+            {
+              path: 'maven4/pom.xml',
+              createIfMissing: false,
+              cachedFileContents: buildGitHubFileContent(
+                fixturesPath,
+                'maven4/pom.xml'
+              ),
+              updater: customUpdater,
+            },
+          ],
+          notes: '### Dependencies\n\n* Updated foo to v3',
+        }),
+      ];
+      stubFilesFromFixtures({
+        sandbox,
+        github,
+        fixturePath: fixturesPath,
+        files: [
+          'maven1/pom.xml',
+          'maven2/pom.xml',
+          'maven3/pom.xml',
+          'maven4/pom.xml',
+        ],
+        flatten: false,
+        targetBranch: 'main',
+      });
+      sandbox
+        .stub(github, 'findFilesByFilenameAndRef')
+        .withArgs('pom.xml', 'main')
+        .resolves([
+          'maven1/pom.xml',
+          'maven2/pom.xml',
+          'maven3/pom.xml',
+          'maven4/pom.xml',
+        ]);
+      const newCandidates = await plugin.run(candidates);
+      expect(newCandidates).length(1);
+      safeSnapshot(newCandidates[0].pullRequest.body.toString());
+      expect(newCandidates[0].pullRequest.body.releaseData).length(2);
+      assertHasUpdates(
+        newCandidates[0].pullRequest.updates,
+        'maven4/pom.xml',
+        RawContent,
+        PomXml
+      );
+      assertHasUpdates(
+        newCandidates[0].pullRequest.updates,
+        'maven3/pom.xml',
+        PomXml,
+        PomXml
+      );
+    });
     it('walks dependency tree and updates previously untouched packages', async () => {
       const candidates: CandidateReleasePullRequest[] = [
-        buildMockCandidatePullRequest('maven1', 'maven', '1.1.2', 'maven1', [
-          buildMockPackageUpdate('maven1/pom.xml', 'maven1/pom.xml', '1.1.2'),
-        ]),
+        buildMockCandidatePullRequest('maven1', 'maven', '1.1.2', {
+          component: 'maven1',
+          updates: [
+            buildMockPackageUpdate('maven1/pom.xml', 'maven1/pom.xml', '1.1.2'),
+          ],
+        }),
       ];
       stubFilesFromFixtures({
         sandbox,
@@ -168,6 +246,30 @@ describe('MavenWorkspace plugin', () => {
       expect(newCandidates[0].pullRequest.body.releaseData).length(4);
     });
     it('skips pom files not configured for release', async () => {
+      plugin = new MavenWorkspace(
+        github,
+        'main',
+        {
+          bom: {
+            releaseType: 'maven',
+          },
+          maven1: {
+            releaseType: 'maven',
+          },
+          maven2: {
+            releaseType: 'maven',
+          },
+          maven3: {
+            releaseType: 'maven',
+          },
+          maven4: {
+            releaseType: 'maven',
+          },
+        },
+        {
+          considerAllArtifacts: false,
+        }
+      );
       sandbox
         .stub(github, 'findFilesByFilenameAndRef')
         .withArgs('pom.xml', 'main')
@@ -179,9 +281,12 @@ describe('MavenWorkspace plugin', () => {
           'extra/pom.xml',
         ]);
       const candidates: CandidateReleasePullRequest[] = [
-        buildMockCandidatePullRequest('maven1', 'maven', '1.1.2', 'maven1', [
-          buildMockPackageUpdate('maven1/pom.xml', 'maven1/pom.xml', '1.1.2'),
-        ]),
+        buildMockCandidatePullRequest('maven1', 'maven', '1.1.2', {
+          component: 'maven1',
+          updates: [
+            buildMockPackageUpdate('maven1/pom.xml', 'maven1/pom.xml', '1.1.2'),
+          ],
+        }),
       ];
       stubFilesFromFixtures({
         sandbox,
@@ -201,6 +306,103 @@ describe('MavenWorkspace plugin', () => {
       safeSnapshot(newCandidates[0].pullRequest.body.toString());
       expect(newCandidates[0].pullRequest.body.releaseData).length(4);
     });
+    it('can consider all artifacts', async () => {
+      plugin = new MavenWorkspace(
+        github,
+        'main',
+        {
+          bom: {
+            component: 'my-bom',
+            releaseType: 'java-yoshi',
+          },
+          multi1: {
+            component: 'multi1',
+            releaseType: 'java-yoshi',
+          },
+          multi2: {
+            component: 'multi2',
+            releaseType: 'java-yoshi',
+          },
+        },
+        {
+          considerAllArtifacts: true,
+        }
+      );
+      const candidates: CandidateReleasePullRequest[] = [
+        buildMockCandidatePullRequest('multi1', 'java-yoshi', '1.1.2', {
+          component: 'multi1',
+          updates: [
+            buildMockPackageUpdate('multi1/pom.xml', 'multi1/pom.xml', '1.1.2'),
+            buildMockPackageUpdate(
+              'multi1/bom/pom.xml',
+              'multi1/bom/pom.xml',
+              '1.1.2'
+            ),
+            buildMockPackageUpdate(
+              'multi1/primary/pom.xml',
+              'multi1/primary/pom.xml',
+              '1.1.2'
+            ),
+            buildMockPackageUpdate(
+              'multi1/sub1/pom.xml',
+              'multi1/sub1/pom.xml',
+              '2.2.3'
+            ),
+            buildMockPackageUpdate(
+              'multi1/sub2/pom.xml',
+              'multi1/sub2/pom.xml',
+              '3.3.4'
+            ),
+          ],
+        }),
+      ];
+      stubFilesFromFixtures({
+        sandbox,
+        github,
+        fixturePath: fixturesPath,
+        files: [
+          'bom/pom.xml',
+          'multi1/pom.xml',
+          'multi1/bom/pom.xml',
+          'multi1/primary/pom.xml',
+          'multi1/sub1/pom.xml',
+          'multi1/sub2/pom.xml',
+          'multi2/pom.xml',
+          'multi2/bom/pom.xml',
+          'multi2/primary/pom.xml',
+          'multi2/sub1/pom.xml',
+          'multi2/sub2/pom.xml',
+        ],
+        flatten: false,
+        targetBranch: 'main',
+      });
+      sandbox
+        .stub(github, 'findFilesByFilenameAndRef')
+        .withArgs('pom.xml', 'main')
+        .resolves([
+          'bom/pom.xml',
+          'multi1/pom.xml',
+          'multi1/bom/pom.xml',
+          'multi1/primary/pom.xml',
+          'multi1/sub1/pom.xml',
+          'multi1/sub2/pom.xml',
+          'multi2/pom.xml',
+          'multi2/bom/pom.xml',
+          'multi2/primary/pom.xml',
+          'multi2/sub1/pom.xml',
+          'multi2/sub2/pom.xml',
+        ]);
+      const newCandidates = await plugin.run(candidates);
+      expect(newCandidates).length(1);
+      expect(newCandidates[0].pullRequest.body.releaseData).length(2);
+      safeSnapshot(newCandidates[0].pullRequest.body.toString());
+      const bomUpdate = assertHasUpdate(
+        newCandidates[0].pullRequest.updates,
+        'bom/pom.xml',
+        PomXml
+      );
+      safeSnapshot(await renderUpdate(github, bomUpdate));
+    });
   });
 });
 
@@ -216,4 +418,10 @@ function buildMockPackageUpdate(
     cachedFileContents,
     updater: new PomXml(Version.parse(newVersionString)),
   };
+}
+
+async function renderUpdate(github: GitHub, update: Update): Promise<string> {
+  const fileContents =
+    update.cachedFileContents || (await github.getFileContents(update.path));
+  return update.updater.updateContent(fileContents.parsedContent);
 }