+++ test/strategies/base.ts
@@ -74,6 +74,24 @@ describe('Strategy', () => {
       expect(pullRequest?.version?.toString()).to.eql('2.3.4');
       snapshot(dateSafe(pullRequest!.body.toString()));
     });
+    it('allows overriding initial version in base constructor', async () => {
+      const strategy = new TestStrategy({
+        targetBranch: 'main',
+        github,
+        component: 'google-cloud-automl',
+        initialVersion: '0.1.0',
+      });
+      const commits = [
+        {
+          sha: 'abc123',
+          message: 'feat: initial commit',
+        },
+      ];
+      const pullRequest = await strategy.buildReleasePullRequest(commits);
+      expect(pullRequest).to.not.be.undefined;
+      expect(pullRequest?.version?.toString()).to.eql('0.1.0');
+      snapshot(dateSafe(pullRequest!.body.toString()));
+    });
     it('updates extra files', async () => {
       const strategy = new TestStrategy({
         targetBranch: 'main',
@@ -246,6 +264,20 @@ describe('Strategy', () => {
           );
         }
       }
+    });
+    it('handles extra labels', async () => {
+      const strategy = new TestStrategy({
+        targetBranch: 'main',
+        github,
+        component: 'google-cloud-automl',
+        extraLabels: ['foo', 'bar'],
+      });
+      const pullRequest = await strategy.buildReleasePullRequest(
+        [{sha: 'aaa', message: 'fix: a bugfix'}],
+        undefined
+      );
+      expect(pullRequest).to.exist;
+      expect(pullRequest?.labels).to.eql(['foo', 'bar']);
     });
   });
   describe('buildRelease', () => {